+++ src/components/job-form.tsx
@@ -32,6 +32,7 @@ import { FaFileUpload } from 'react-icons/fa';
 import { Switch } from './ui/switch';
 import { Label } from './ui/label';
 import dynamic from 'next/dynamic';
+import { uploadFileAction } from '@/actions/upload-to-cdn';
 
 const DynamicLineDrawingAnimation = dynamic(
   () => import('./gmaps-autosuggest'),
@@ -86,16 +87,12 @@ const PostJobForm = () => {
       const uniqueFileName = `${Date.now()}-${file.name}`;
       formData.append('uniqueFileName', uniqueFileName);
 
-      const res = await fetch(`/api/upload-to-cdn`, {
-        method: 'POST',
-        body: formData,
-      });
-
-      if (!res.ok) {
+      const res = await uploadFileAction(formData);
+      if (!res) {
         throw new Error('Failed to upload image');
       }
 
-      const uploadRes = await res.json();
+      const uploadRes = res;
       return uploadRes.url;
     } catch (error) {
       console.error('Image upload failed:', error);
